pyproject.toml
[project]
name = "mlops-zoomcamp-epam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.2",
    "boto3>=1.38.46",
    "fastparquet>=2024.11.0",
    "hyperopt>=0.2.7",
    "matplotlib>=3.10.3",
    "mlflow>=3.1.1",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "xgboost>=3.0.2",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5", "uv>=0.7.13"]

# For mlflow docker server, though it does not change much (image still big), won't use it for now.
mlflow-server = ["mlflow>=3.1.1", "pandas>=2.3.0", "setuptools>=80.9.0"]

airflow-server = [
    "xgboost>=3.0.2",
    "pandas>=2.3.0",
    "hyperopt>=0.2.7",
    "scikit-learn>=1.7.0",
]

[tool.setuptools.packages.find]
include = ["dags*"]
namespaces = false

[tool.ruff.lint]
select = ["ALL"]
extend-select = ["I"]